% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiteam.R
\name{ggsave_for_docs}
\alias{ggsave_for_docs}
\title{Plot save function}
\usage{
ggsave_for_docs(fig, plot.dir, plot.name, height, width, save.fig,
  save.format)
}
\arguments{
\item{fig}{ggplot object}

\item{plot.dir}{string, name of directory for output file}

\item{plot.name}{string, desired name for plot file}

\item{height}{numeric, height of plot in inches}

\item{width}{numeric, width of plot in in inches}

\item{save.fig}{logical, whether plot figure will be saved or not}

\item{save.format}{string, "svg", "png", "emf"}
}
\description{
This function will display and save a ggplot figure.
}
\examples{
ggsave_for_docs(fig = my_plot,
 plot.dir = "plot_directory/",
 plot.name = "my_plot",
 height = 4,
 width = 6,
 save.fig = T,
 save.format = "png")
}
