% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiteam.R
\docType{data}
\name{iteam.colors}
\alias{iteam.colors}
\title{Baltimore i-team colors}
\format{An object of class \code{character} of length 6.}
\usage{
iteam.colors
}
\description{
Baltimore i-team colors
}
\keyword{datasets}
