% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiteam.R
\name{scale_color_continuous_iteam}
\alias{scale_color_continuous_iteam}
\title{Baltimore i-team's ggplot2 colors for continuous color aesthetics}
\usage{
scale_color_continuous_iteam()
}
\description{
Works the same as scale_color_continuous for ggplot plots.
}
\examples{
library(tidyverse)
mtcars \%>\%
  ggplot(aes(wt, hp, color = disp)) +
  geom_point() +
  scale_color_continuous_iteam()
}
