% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiteam.R
\name{scale_color_discrete_iteam}
\alias{scale_color_discrete_iteam}
\title{Baltimore i-team's ggplot2 colors for discrete aesthetics}
\usage{
scale_color_discrete_iteam()
}
\description{
Works the same as scale_color_discrete for ggplot plots.
}
\examples{
library(tidyverse)
mtcars \%>\%
  ggplot(aes(wt, mpg, color = factor(cyl))) +
  geom_point() +
  scale_color_discrete_iteam()
}
