% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiteam.R
\name{scale_fill_discrete_iteam}
\alias{scale_fill_discrete_iteam}
\title{Baltimore i-team's ggplot2 colors for discrete aesthetics}
\usage{
scale_fill_discrete_iteam()
}
\description{
Works the same as scale_fill_discrete for ggplot plots.
}
\examples{
library(tidyverse)
mtcars \%>\%
  ggplot(aes(x = am, y = ..count.., fill = factor(cyl))) +
  geom_bar() +
  scale_fill_discrete_iteam()
}
