% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiteam.R
\name{theme_iteam_google_docs}
\alias{theme_iteam_google_docs}
\title{Baltimore i-team's ggplot2 Theme for Google Docs}
\usage{
theme_iteam_google_docs()
}
\description{
This function will theme a ggplot2 theme for Google Docs using the i-team's
colors and design choices. It works the same as any other ggplot2 theme.
}
\examples{
library(tidyverse)
mtcars \%>\%
 ggplot(aes(x = wt, y = mpg, color = disp)) +
 geom_point(size = 5) +
 theme_iteam_google_docs()
}
