% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiteam.R
\name{theme_iteam_presentations}
\alias{theme_iteam_presentations}
\title{Baltimore i-team's ggplot2 Theme for Slide Decks}
\usage{
theme_iteam_presentations()
}
\description{
This function will theme a ggplot2 theme for slide decks using the i-team's
colors and design choices. It works the same as any other ggplot2 theme.
}
\examples{
mtcars \%>\%
 ggplot(aes(x = wt, y = mpg, color = disp)) +
 geom_point(size = 5) +
 theme_iteam_presentations()
}
